HMAC in recv implementation and working!
Line 65: Created new variable called msg_length to help break up the
concatenated message and hash
Line 67: comb_data variable created for decrypted message + plaintext
string
Line 68 and 69: Breaking it up into the message and hash. 32 used as the
size of the decrypted hmac is 32 bytes long (SHA256)
Line 70: Hash created to check hash of received message
Lines 76-80: Checking to see if the hash of the received message is the
same as the hash of the sent message. Actions done by code according to
if it has been altered or not