Initial replay prevention implementation
In comms.py
Lines 42 -52: New lines to create seed for 'counters' in sending and
receiving packets. Initially each bot makes their own random int for one
counter and sends it over for the other bot to use as a seed for the
second counter. Planned future implementation commented in lines 42 and
43 as discussed. Currently received int not generated by connected bot.

lcg generate: Made variables local to the method and cleaned up
indentation. To be fixed. LCG will be used as predictable prng for
counter