Verify Signature Partial Implementation
In files.py:
Lines 2-4: imported relevant libraries
Lines 39-47: code for verifying signature added. Mostly the one shown in
Armie's doco but with changes to suit our variables.

NEED TO DO: Code does not work. Hashed_file needs to be the original
message for it to be properly verified. Need to find way to get original
file to be hashed here