Signature verification working
Hardcoded original hello.fbi file to open in files.py to get the
original file to verify the signature. Hardcoded because this is the
only file we are testing for verification

Added the public keys and private keys of the master bot to ensure
everyone has the same keys. Commented out function call to generate keys
in files.py to prevent someone from having different keys in their local
repo

Added hello.fbi.signed file to ensure everyone has the same signature