Encryption working in cbc
Inserted a method to insert padding to the message for encryption. Given
in the labs and readily available.

Altered encryption code to suit cbc. Code returns no errors and does not
crash when executed. Commented out some of the cfb lines in case we
decide to opt for that mode of operation. Commented out print lines for
cleaner output in console (can be uncommented if needed but recommended
to use debugger instead)