Code added to write aes key into file
Commented out line 61. Will be removed

Lines 49-58: New code to write the aes key into the file. Lines 49 and
50 encrypt it in rsa. Changed the variable name in lines 49-50 to
indicate this is the aes key, not th rsa one

File currently saves in pastebot.net and prints out if file written
successfully.

NOTE: The variable encrypt_aes_key is a tuple, hence to get the bytes
stream the first element of the tuple was taken and written into the
file