Improved signature code
In old code, the signed file contained just the signature and the
original hello.fbi file was hardcoded to be opened for verification. In
this commit:

For master_sign.py:
Original file written into signed file to accomodate this.
Few cleanups done as well removing redundant imports and comments in old
lines 4 and 32.

For files.py:
Comments added to verify_file.
Slight coding changes to extract the original contents in line 84 and
the signature in line 89.